Adding support for voxel scales other than 1
In most cases you probably want a voxel scale to be larger than 1m, as that's generally SMALLER than a character height, and since doorways need to fit in a single voxel.... voxels need to be larger than the characters probably. This should resolve this.

See comments on the voxelscale property in DungeonGenerator.cs  on how to use this